+++ packages/plugin-collapsible-list/web/src/consts.ts
@@ -6,6 +6,8 @@ export const collapsibleModals = {
 
 export const collapsibleStateManagerPlugin = new PluginKey('collapsibleStateManager');
 
+export const collapsibleItemDropHandlerPlugin = new PluginKey('collapsibleItemDropHandler');
+
 export const COLLAPSIBLE_EXPAND_STATE = {
   EXPANDED: 'expanded',
   FIRST: 'first_expanded',