+++ packages/plugin-divider/web/src/getAddButtons.tsx
@@ -2,6 +2,7 @@ import InsertPluginIcon from './icons/InsertPluginIcon';
 import { INSERT_PLUGIN_BUTTONS, TOOLBARS } from 'wix-rich-content-editor-common';
 import type { AddButton } from 'ricos-types';
 import { DIVIDER_TYPE } from './types';
+import { DividerData_Width } from 'ricos-schema';
 
 export const getAddButtons = (config, services): AddButton[] => {
   return [
@@ -13,7 +14,7 @@ export const getAddButtons = (config, services): AddButton[] => {
       tooltip: 'DividerPlugin_InsertButton_Tooltip',
       toolbars: [TOOLBARS.MOBILE, TOOLBARS.FOOTER, TOOLBARS.SIDE],
       command: editorCommands => {
-        editorCommands.insertBlockWithBlankLines(DIVIDER_TYPE, {});
+        editorCommands.insertBlockWithBlankLines(DIVIDER_TYPE, { width: DividerData_Width.LARGE });
         return true;
       },
       menuConfig: {